#1 管理者側　店舗レビュー一覧・削除機能　追加

+++ kadai_002/nagoyameshi/routes/web.php
@@ -45,6 +45,7 @@ Route::group(['prefix' => 'dashboard', 'as' => 'dashboard.'], function () {
     Route::post('login', [LoginController::class, 'login'])->name('login');
     Route::resource('categories', CategoryController::class)->middleware('auth:admins');
     Route::resource('restaurants', RestaurantController::class)->middleware('auth:admins');
+    Route::get('restaurants/{restaurant}/reviews', [RestaurantController::class, 'reviews'])->middleware('auth:admins')->name('restaurants.reviews');
     Route::get('company', [CompanyController::class, 'index'])->middleware('auth:admins')->name('company.index');
     Route::put('company/{company}', [CompanyController::class, 'update'])->middleware('auth:admins')->name('company.update');
     Route::get('users', [UserController::class, 'index'])->middleware('auth:admins')->name('users.index');
@@ -68,7 +69,7 @@ Route::post('reviews', [ReviewController::class, 'store'])->middleware('auth')->
 Route::get('reviews/{review}', [ReviewController::class, 'show'])->name('reviews.show');
 Route::get('reviews/{review}/edit/{restaurant}', [ReviewController::class, 'edit'])->middleware('auth')->name('reviews.edit');
 Route::put('reviews/{review}', [ReviewController::class, 'update'])->middleware('auth')->name('reviews.update');
-Route::delete('/reviews/{review}', [ReviewController::class, 'destroy'])->middleware('auth')->name('reviews.destroy');
+Route::delete('/reviews/{review}', [ReviewController::class, 'destroy'])->middleware(['auth', 'auth:admins'])->name('reviews.destroy');
 
 Route::controller(mainUserController::class)->group(function () {
     Route::get('users/mypage', 'mypage')->name('mypage');
